--- src/Controls/samples/Controls.Sample.UITests/Elements/NavBarTranslucentGalleryPage.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.Maui.Controls.Internals;
-using Microsoft.Maui.Controls.PlatformConfiguration;
-using Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;
-
-namespace Maui.Controls.Sample
-{
-	[Preserve(AllMembers = true)]
-	public class NavBarTranslucentGalleryPage : ContentViewGalleryPage
-	{
-		public NavBarTranslucentGalleryPage()
-		{
-			Content = new NavBarTranslucentPage();
-		}
-	}
-}
