+++ src/Controls/samples/Controls.Sample.UITests/Issues/Issue17022.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              AutomationId="NavBarTranslucentPage"
-             x:Class="Maui.Controls.Sample.NavBarTranslucentPage"
+             x:Class="Maui.Controls.Sample.Issues.Issue17022"
              xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
             ios:Page.UseSafeArea="False"
             ios:NavigationPage.IsNavigationBarTranslucent="True"
@@ -11,4 +11,4 @@
         <BoxView Background="Green" Grid.Row="0" AutomationId="TopBoxView"/>
         <Label Margin="20" Text="Green boxview should be behind navbar and touching very top of screen" Grid.Row="1" />
     </Grid>
-</ContentView>
+</ContentPage>
